Clean up utils and unify job constants usage
