--- nvflare/fuel/utils/check_utils.py
@@ -1,23 +1,0 @@
-# Copyright (c) 2021-2022, NVIDIA CORPORATION.  All rights reserved.
-#
-# Licensed under the Apache License, Version 2.0 (the "License");
-# you may not use this file except in compliance with the License.
-# You may obtain a copy of the License at
-#
-#     http://www.apache.org/licenses/LICENSE-2.0
-#
-# Unless required by applicable law or agreed to in writing, software
-# distributed under the License is distributed on an "AS IS" BASIS,
-# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-# See the License for the specific language governing permissions and
-# limitations under the License.
-
-from typing import Any
-
-
-def check_non_neg_int(data: Any, name: str):
-    if not isinstance(data, int):
-        raise ValueError(f"{name} must be int but got {type(data)}")
-
-    if data < 0:
-        raise ValueError(f"{name} must be greater than or equal to 0.")
