+++ tests/unit_test/fuel/utils/fobs/fobs_test.py
@@ -11,9 +11,10 @@
 # WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 # See the License for the specific language governing permissions and
 # limitations under the License.
+
 import queue
 from datetime import datetime
-from typing import Any, Type
+from typing import Any
 
 import pytest
