+++ tests/unit_test/fuel/utils/zip_utils_test.py
@@ -19,8 +19,7 @@ from pathlib import Path
 
 import pytest
 
-from nvflare.fuel.hci.zip_utils import (
-    convert_legacy_zip,
+from nvflare.fuel.utils.zip_utils import (
     get_all_file_paths,
     normpath_for_zip,
     remove_leading_dotdot,
@@ -28,79 +27,6 @@ from nvflare.fuel.hci.zip_utils import (
     unzip_all_from_bytes,
     zip_directory_to_bytes,
 )
-
-
-def create_fake_job(temp_dir, job_name):
-    root_dir = Path(temp_dir) / job_name
-    os.makedirs(root_dir)
-    os.mkdir(root_dir / "fed_avg_app")
-    os.mkdir(root_dir / "fed_avg_app" / "config")
-    open(root_dir / "fed_avg_app" / "config" / "config_fed_server.json", "w").close()
-    open(root_dir / "fed_avg_app" / "config" / "config_fed_client.json", "w").close()
-    os.mkdir(root_dir / "fed_avg_app" / "custom")
-    open(root_dir / "fed_avg_app" / "custom" / "c1.py", "w").close()
-    open(root_dir / "fed_avg_app" / "custom" / "c2.py", "w").close()
-    with open(root_dir / "meta.json", "w") as f:
-        f.write("{}")
-
-
-@pytest.fixture()
-def create_fake_app_dir():
-    """
-    fed_avg_app/
-        config/
-            config_fed_server.json
-            config_fed_client.json
-        custom/
-            c1.py
-            c2.py
-    expect result:
-        fed_avg_app/
-            fed_avg_app/
-                config/
-                    config_fed_server.json
-                    config_fed_client.json
-                custom/
-                    c1.py
-                    c2.py
-            meta.json
-    """
-    temp_dir = tempfile.mkdtemp()
-    app_name = "fed_avg_app"
-    root_dir = Path(temp_dir) / app_name
-    os.makedirs(root_dir)
-    os.mkdir(root_dir / "config")
-    open(root_dir / "config" / "config_fed_server.json", "w").close()
-    open(root_dir / "config" / "config_fed_client.json", "w").close()
-    os.mkdir(root_dir / "custom")
-    open(root_dir / "custom" / "c1.py", "w").close()
-    open(root_dir / "custom" / "c2.py", "w").close()
-
-    temp_dir2 = tempfile.mkdtemp()
-    create_fake_job(temp_dir2, "fed_avg_app")
-    yield temp_dir, app_name, temp_dir2
-    shutil.rmtree(temp_dir)
-    shutil.rmtree(temp_dir2)
-
-
-@pytest.fixture()
-def create_fake_job_dir():
-    """
-    fed_avg/
-        fed_avg_app/
-            config/
-                config_fed_server.json
-                config_fed_client.json
-            custom/
-                c1.py
-                c2.py
-        meta.json
-    """
-    temp_dir = tempfile.mkdtemp()
-    job_name = "fed_avg"
-    create_fake_job(temp_dir, job_name)
-    yield temp_dir, job_name
-    shutil.rmtree(temp_dir)
 
 
 @pytest.fixture()
@@ -142,7 +68,7 @@ def create_fake_dir():
     shutil.rmtree(temp_dir)
 
 
-class TestSecurityUtils:
+class TestZipUtils:
     @pytest.mark.parametrize(
         "path, output",
         [
@@ -204,30 +130,3 @@ class TestSecurityUtils:
             assert i[1] == j[1]
             assert i[2] == j[2]
         shutil.rmtree(temp_dir)
-
-    def test_convert_legacy_zip_job(self, create_fake_job_dir):
-        tmp_dir, job_name = create_fake_job_dir
-        zip_data = zip_directory_to_bytes(root_dir=tmp_dir, folder_name=job_name)
-        new_bytes = convert_legacy_zip(zip_data)
-
-        output_tmp_dir = tempfile.mkdtemp()
-        unzip_all_from_bytes(new_bytes, output_dir_name=output_tmp_dir)
-        # stays the same
-        for i, j in zip(os.walk(tmp_dir), os.walk(output_tmp_dir)):
-            assert i[1] == j[1]
-            assert i[2] == j[2]
-
-        shutil.rmtree(output_tmp_dir)
-
-    def test_convert_legacy_zip_app(self, create_fake_app_dir):
-        tmp_dir, app_name, tmp_dir_with_job = create_fake_app_dir
-        zip_data = zip_directory_to_bytes(root_dir=tmp_dir, folder_name=app_name)
-        new_bytes = convert_legacy_zip(zip_data)
-
-        output_tmp_dir = tempfile.mkdtemp()
-        unzip_all_from_bytes(new_bytes, output_dir_name=output_tmp_dir)
-        for i, j in zip(os.walk(tmp_dir_with_job), os.walk(output_tmp_dir)):
-            assert i[1] == j[1]
-            assert i[2] == j[2]
-
-        shutil.rmtree(output_tmp_dir)
